requirements.txt
boto3==1.34.0
botocore==1.34.0
pyyaml==6.0.1
